+++ pykeg/core/management/commands/run_workers.py
@@ -1,4 +1,7 @@
 import os
+import sys
+
+from django.conf import settings
 
 from pykeg.core.management.commands.common import RunnerCommand
 
@@ -8,24 +11,13 @@ class Command(RunnerCommand):
     pidfile_name = "kegbot_run_workers.pid"
 
     def get_commands(self, options):
-        default_log = stats_log = beat_log = ""
+        default_log = ""
         logs_dir = options.get("logs_dir")
-
         if logs_dir:
-            default_log = ' --logfile="{}"'.format(os.path.join(logs_dir, "celery_default.log"))
-            stats_log = ' --logfile="{}"'.format(os.path.join(logs_dir, "celery_stats.log"))
-        ret = []
+            default_log = ' --logfile="{}"'.format(os.path.join(logs_dir, "workers.log"))
 
-        base_cmd = "celery -A pykeg worker -l info "
-
-        ret.append(
-            ("celery_default", base_cmd + '-Q default --hostname="default@%h"' + default_log)
-        )
-        ret.append(
-            (
-                "celery_stats",
-                base_cmd + '-Q stats --concurrency=1 --hostname="stats@%h"' + stats_log,
-            )
-        )
-
+        queue_names = " ".join(settings.RQ_QUEUES.keys())
+        ret = [
+            ("rq", f"{sys.argv[0]} rqworker {queue_names}{default_log} -v 3"),
+        ]
         return ret